Do not override UiKit message surface styles with message content body styles
