+++ apps/meteor/client/components/message/MessageContentBody.tsx
@@ -1,112 +1,16 @@
 import { css } from '@rocket.chat/css-in-js';
 import { MessageBody, Box, Palette } from '@rocket.chat/fuselage';
-import type { UserMention, ChannelMention } from '@rocket.chat/gazzodown';
-import { MarkupInteractionContext, Markup } from '@rocket.chat/gazzodown';
-import { escapeRegExp } from '@rocket.chat/string-helpers';
-import { useLayout, useUserPreference } from '@rocket.chat/ui-contexts';
-import { FlowRouter } from 'meteor/kadira:flow-router';
-import type { ReactElement, UIEvent } from 'react';
-import React, { useCallback, useMemo } from 'react';
+import React from 'react';
 
-import { emoji } from '../../../app/emoji/client';
 import type { MessageWithMdEnforced } from '../../lib/parseMessageTextToAstMarkdown';
-import { fireGlobalEvent } from '../../lib/utils/fireGlobalEvent';
-import { useChat } from '../../views/room/contexts/ChatContext';
-import { useGoToRoom } from '../../views/room/hooks/useGoToRoom';
-import { useMessageListHighlights } from './list/MessageListContext';
+import GazzodownText from '../GazzodownText';
 
 type MessageContentBodyProps = Pick<MessageWithMdEnforced, 'mentions' | 'channels' | 'md'> & {
 	searchText?: string;
 };
 
-const detectEmoji = (text: string): { name: string; className: string; image?: string; content: string }[] => {
-	const html = Object.values(emoji.packages)
-		.reverse()
-		.reduce((html, { render }) => render(html), text);
-	const div = document.createElement('div');
-	div.innerHTML = html;
-	return Array.from(div.querySelectorAll('span')).map((span) => ({
-		name: span.title,
-		className: span.className,
-		image: span.style.backgroundImage || undefined,
-		content: span.innerText,
-	}));
-};
-
-const MessageContentBody = ({ mentions, channels, md, searchText }: MessageContentBodyProps): ReactElement => {
-	const highlights = useMessageListHighlights();
-	const highlightRegex = useMemo(() => {
-		if (!highlights || !highlights.length) {
-			return;
-		}
-
-		const alternatives = highlights.map(({ highlight }) => escapeRegExp(highlight)).join('|');
-		const expression = `(?=^|\\b|[\\s\\n\\r\\t.,،'\\\"\\+!?:-])(${alternatives})(?=$|\\b|[\\s\\n\\r\\t.,،'\\\"\\+!?:-])`;
-
-		return (): RegExp => new RegExp(expression, 'gmi');
-	}, [highlights]);
-
-	const markRegex = useMemo(() => {
-		if (!searchText) {
-			return;
-		}
-
-		return (): RegExp => new RegExp(`(${searchText})(?![^<]*>)`, 'gi');
-	}, [searchText]);
-
-	const resolveUserMention = useCallback(
-		(mention: string) => {
-			if (mention === 'all' || mention === 'here') {
-				return undefined;
-			}
-
-			const filterUser = ({ username, type }: UserMention) => (!type || type === 'user') && username === mention;
-			const filterTeam = ({ name, type }: UserMention) => type === 'team' && name === mention;
-
-			return mentions?.find((mention) => filterUser(mention) || filterTeam(mention));
-		},
-		[mentions],
-	);
-
-	const chat = useChat();
-
-	const goToRoom = useGoToRoom();
-
-	const onUserMentionClick = useCallback(
-		({ username }: UserMention) => {
-			if (!username) {
-				return;
-			}
-
-			return (event: UIEvent): void => {
-				event.stopPropagation();
-				chat?.userCard.open(username)(event);
-			};
-		},
-		[chat?.userCard],
-	);
-
-	const resolveChannelMention = useCallback((mention: string) => channels?.find(({ name }) => name === mention), [channels]);
-
-	const { isEmbedded } = useLayout();
-
-	const onChannelMentionClick = useCallback(
-		({ _id: rid }: ChannelMention) =>
-			(event: UIEvent): void => {
-				if (isEmbedded) {
-					fireGlobalEvent('click-mention-link', {
-						path: FlowRouter.path('channel', { name: rid }),
-						channel: rid,
-					});
-				}
-
-				event.stopPropagation();
-				goToRoom(rid);
-			},
-		[isEmbedded, goToRoom],
-	);
-
-	// TODO:  this style should go to Fuselage <MessageBody> repository
+const MessageContentBody = ({ mentions, channels, md, searchText }: MessageContentBodyProps) => {
+	// TODO: this style should go to Fuselage <MessageBody> repository
 	const messageBodyAdditionalStyles = css`
 		> blockquote {
 			padding-inline: 8px;
@@ -147,25 +51,10 @@ const MessageContentBody = ({ mentions, channels, md, searchText }: MessageConte
 		}
 	`;
 
-	const convertAsciiToEmoji = useUserPreference<boolean>('convertAsciiEmoji', true);
-
 	return (
 		<MessageBody data-qa-type='message-body'>
 			<Box className={messageBodyAdditionalStyles}>
-				<MarkupInteractionContext.Provider
-					value={{
-						detectEmoji,
-						highlightRegex,
-						markRegex,
-						resolveUserMention,
-						onUserMentionClick,
-						resolveChannelMention,
-						onChannelMentionClick,
-						convertAsciiToEmoji,
-					}}
-				>
-					<Markup tokens={md} />
-				</MarkupInteractionContext.Provider>
+				<GazzodownText tokens={md} channels={channels} mentions={mentions} searchText={searchText} />
 			</Box>
 		</MessageBody>
 	);
