+++ apps/meteor/client/components/message/content/uikit/ParsedText.tsx
@@ -1,13 +1,25 @@
+import type { IRoom } from '@rocket.chat/core-typings';
 import { parse } from '@rocket.chat/message-parser';
-import type { ReactElement } from 'react';
 import React, { memo, useMemo } from 'react';
 
-import MessageContentBody from '../../MessageContentBody';
+import GazzodownText from '../../../GazzodownText';
 
-const ParsedText = ({ text }: { text: string }): ReactElement | null => {
-	const md = useMemo(() => parse(text, { emoticons: true }), [text]);
+type ParsedTextProps = {
+	text: string;
+	mentions?: {
+		type: 'user' | 'team';
+		_id: string;
+		username?: string;
+		name?: string;
+	}[];
+	channels?: Pick<IRoom, '_id' | 'name'>[];
+	searchText?: string;
+};
 
-	return <MessageContentBody md={md} />;
+const ParsedText = ({ text, mentions, channels, searchText }: ParsedTextProps) => {
+	const tokens = useMemo(() => parse(text, { emoticons: true }), [text]);
+
+	return <GazzodownText tokens={tokens} mentions={mentions} channels={channels} searchText={searchText} />;
 };
 
 export default memo(ParsedText);
