+++ apps/meteor/client/providers/RouterProvider.tsx
@@ -1,6 +1,7 @@
 import type { RouterContextValue } from '@rocket.chat/ui-contexts';
 import { RouterContext } from '@rocket.chat/ui-contexts';
 import { FlowRouter } from 'meteor/kadira:flow-router';
+import { Tracker } from 'meteor/tracker';
 import type { FC } from 'react';
 import React from 'react';
 
@@ -51,6 +52,9 @@ const queryCurrentRoute = (): ReturnType<RouterContextValue['queryCurrentRoute']
 		return [route?.name, params, queryParams, route?.group?.name];
 	});
 
+const getRoutePath = (name: string, parameters?: Record<string, string>, queryStringParameters?: Record<string, string>) =>
+	Tracker.nonreactive(() => FlowRouter.path(name, parameters, queryStringParameters));
+
 const contextValue = {
 	queryRoutePath,
 	queryRouteUrl,
@@ -59,6 +63,7 @@ const contextValue = {
 	queryRouteParameter,
 	queryQueryStringParameter,
 	queryCurrentRoute,
+	getRoutePath,
 };
 
 const RouterProvider: FC = ({ children }) => <RouterContext.Provider children={children} value={contextValue} />;
